Move client config to its own package
